==> 02-opennemco2eqcalculator.ipynb <==
import pandas as pd

df = pd.read_csv('./opennem/yearly/2020-2021-opennem-nsw.csv')
df

df.info()

df['Total Generation'] = df['Pumps -  GWh'] + df['Coal (Black) -  GWh'] + df['Distillate -  GWh'] + df['Gas (CCGT) -  GWh'] + df['Gas (OCGT) -  GWh'] + df['Hydro -  GWh'] + df['Wind -  GWh'] + df['Solar (Utility) -  GWh'] + df['Solar (Rooftop) -  GWh']

df['Total Generation']

df = df.drop([12])

df['Total Generation']

df['PercentRenewables'] = (df['Hydro -  GWh'] + df['Wind -  GWh'] + df['Solar (Rooftop) -  GWh'] + df['Solar (Utility) -  GWh'])/df['Total Generation']

df['PercentRenewables']

df.info()

unwanted_cols = ['Emissions Intensity - kgCO₂e/MWh','Exports -  GWh','Imports -  GWh','Price - AUD/MWh','Pumps Market Value - AUD','Exports Market Value - AUD','Imports Market Value - AUD','Coal (Black) Market Value - AUD','Distillate Market Value - AUD','Gas (CCGT) Market Value - AUD','Gas (OCGT) Market Value - AUD','Hydro Market Value - AUD','Solar (Rooftop) Market Value - AUD','Solar (Utility) Market Value - AUD','Wind Market Value - AUD'] 
df = df.drop(axis=1, columns=unwanted_cols)

df['TotalCO2Emissions'] = df['Coal (Black) Emissions Vol - tCO₂e'] + df['Gas (CCGT) Emissions Vol - tCO₂e'] + df['Gas (OCGT) Emissions Vol - tCO₂e'] + df['Distillate Emissions Vol - tCO₂e']

df['CarbonIntensity'] = df['TotalCO2Emissions'] / df['Total Generation']

df

# Start Processing VIC data

unwanted_cols = ['Emissions Intensity - kgCO₂e/MWh','Exports -  GWh','Imports -  GWh','Price - AUD/MWh','Exports Market Value - AUD','Imports Market Value - AUD','Gas (OCGT) Market Value - AUD','Hydro Market Value - AUD','Solar (Rooftop) Market Value - AUD','Solar (Utility) Market Value - AUD','Wind Market Value - AUD'] 
df_vic = pd.read_csv('./opennem/yearly/2020-2021-opennem-vic.csv')
df_vic = df_vic.drop(axis=1, columns=unwanted_cols)

df_vic['Total Generation'] =  df_vic['Coal (Brown) -  GWh'] + df_vic['Gas (Steam) -  GWh'] + df_vic['Gas (OCGT) -  GWh'] + df_vic['Battery (Discharging) -  GWh'] + df_vic['Hydro -  GWh'] + df_vic['Wind -  GWh'] + df_vic['Solar (Utility) -  GWh'] + df_vic['Solar (Rooftop) -  GWh']

df_vic['Total Generation']

df_vic['PercentRenewables'] = (df_vic['Battery (Discharging) -  GWh'] + df_vic['Solar (Rooftop) -  GWh'] + df_vic['Solar (Utility) -  GWh'] + df_vic['Hydro -  GWh'] + df_vic['Wind -  GWh'])/df_vic['Total Generation']

df_vic['TotalCO2Emissions'] = df_vic['Coal (Brown) Emissions Vol - tCO₂e'] + df_vic['Gas (OCGT) Emissions Vol - tCO₂e'] + df_vic['Gas (Steam) Emissions Vol - tCO₂e']

df_vic['CarbonIntensity'] = df_vic['TotalCO2Emissions'] / df_vic['Total Generation']

df_vic[['Total Generation','PercentRenewables','CarbonIntensity','TotalCO2Emissions']]

df[['Total Generation','PercentRenewables','CarbonIntensity','TotalCO2Emissions']]

df.to_csv('nsw_co2intensity.csv')
df_vic.to_csv('vic_co2intensity.csv')

df
