==> 03-azvmtdpcalculator.ipynb <==
import pandas as pd

vm_skus = pd.read_csv('./azurevmskus.csv')
vm_skus

vm_skus['memtdp'] = vm_skus['ram']/8 * 3

vm_skus['cputotaltdp'] = vm_skus['corecount']/vm_skus['cpumaxcores'] * vm_skus['cputdp']

vm_skus['gputotaltdp'] = vm_skus['gpucount'] * vm_skus['gputdp']

vm_skus['totaltdp'] = vm_skus['memtdp'] + vm_skus['cputotaltdp'] + vm_skus['gputotaltdp']

vm_skus

vm_skus.to_csv('azurevmskus_withtdp.csv')
