==> 04-azurevmco2eqestimator.ipynb <==
import pandas as pd

nsw_co2 = pd.read_csv('./emissions/nsw_co2intensity.csv')
vic_co2 = pd.read_csv('./emissions/vic_co2intensity.csv')

vm_skus_tdp = pd.read_csv('./azurevms/azurevmskus_withtdp.csv')
vm_skus_tdp

# These are the inputs

desired_sku_list = [{'sku':'Standard_D3_v2','count':4}, {'sku':'Standard_F16s_v2', 'count':2}, {'sku':'Standard_NC12','count': 1}]
default_hours_month = 730.5
start_month = 10
average_util_pct = .30
desired_region = 'vic'
microsoft_pue_constant = 1.125

if(desired_region == 'nsw'):
    carbonintensity = nsw_co2[['date','CarbonIntensity']]
else:
    carbonintensity = vic_co2[['date','CarbonIntensity']]



# Filter skus
skus_list = [x['sku'] for x in desired_sku_list]
is_in_desired_skus = vm_skus_tdp['skuname'].isin(skus_list)

vm_desired_skus = vm_skus_tdp[is_in_desired_skus]
vm_desired_skus

# Append SKU count
vm_desired_skus['skucount'] =  vm_desired_skus['skuname'].map(lambda x: [y['count'] for y in desired_sku_list if y['sku'] == x].pop() )
vm_desired_skus['skucount']

vm_desired_skus

## Calculate total monthly Kwh using formula here (https://devblogs.microsoft.com/sustainable-software/how-can-i-calculate-co2eq-emissions-for-my-azure-vm/)

vm_desired_skus['totalmonthlykwh'] = vm_desired_skus['totaltdp'] * vm_desired_skus['skucount'] * default_hours_month / 1000 * microsoft_pue_constant * average_util_pct
vm_desired_skus['totalmonthlykwh']

carbonintensity['CarbonIntensity']

# Generate CO2eq emissions per month
desired_skus_with_monthly_co2 = vm_desired_skus.apply(lambda x: carbonintensity['CarbonIntensity'].map(lambda y: y * x['totalmonthlykwh']), axis=1)

# Convert to KG
desired_skus_with_monthly_co2 = desired_skus_with_monthly_co2.apply(lambda x: x/1000)

desired_skus_with_monthly_co2

desired_skus_with_monthly_co2['TotalYearlyCO2InKg'] = desired_skus_with_monthly_co2.sum(axis=1)
desired_skus_with_monthly_co2['TotalYearlyCO2InKg']

desired_skus_with_monthly_co2

vm_desired_skus = vm_desired_skus.join(desired_skus_with_monthly_co2)

vm_desired_skus.head(5)
