==> 05-cloudfit.ipynb <==
import pandas as pd

# How to export CloudFit Azure RateCard

Run the kusto query:
```
CarbonEmissionResourceCard_Snapshot |
where Timestamp == '2022-08-01T00:00:00Z' |
join kind=inner MeterMapping_Materialized_Snapshot on MeterGuid |
summarize arg_max(Timestamp,*) by MeterGuid
```

Then export to CSV

az_co2_ratecard = pd.read_csv('./cloudfit/export.csv')

# CloudFit Azure RateCard

These units are units depending on the following classification

| Resource | Common Unit | Definition
|:--:|:--:|:--:|
| Storage | Gb-hour | One GB stored for one hour in Azure
| Compute | Core-hour | One core used for one hour based on the size of VM
| Network | GB | One GB of data transmitted 

az_co2_ratecard[az_co2_ratecard.MeterCategoryName == 'Virtual Machines']

# Constants

hrs_per_month = 730

# Functions to retrieve each component's (compute, network, storage) CO2 emissions

def get_emissions_for_vm_sku_in_kg(vm_sku, region='AU East'):
    return az_co2_ratecard[(az_co2_ratecard.MeterCategoryName == 'Virtual Machines') & (az_co2_ratecard.MeterName == vm_sku) & (az_co2_ratecard.MeterRegionName == region)].iloc[0].EmissionsPerUnitResourceUsage * hrs_per_month * 1000

def get_emissions_for_vm_disk_in_kg(disk_sku, region='AU East'):
    return az_co2_ratecard[(az_co2_ratecard.MeterCategoryName == 'Storage') & (az_co2_ratecard.MeterName == disk_sku) & (az_co2_ratecard.MeterRegionName == region)].iloc[0].EmissionsPerUnitResourceUsage * 1000

def get_emissions_for_vm_network(data_out_in_gb, region='All'):
    return (az_co2_ratecard[(az_co2_ratecard.MeterCategoryName == 'Virtual Network') & (az_co2_ratecard.MeterName == 'Inter-Region Egress') & (az_co2_ratecard.MeterRegionName == region)].iloc[0].EmissionsPerUnitResourceUsage + az_co2_ratecard[(az_co2_ratecard.MeterCategoryName == 'Virtual Network') & (az_co2_ratecard.MeterName == 'Inter-Region Ingress') & (az_co2_ratecard.MeterRegionName == region)].iloc[0].EmissionsPerUnitResourceUsage) * data_out_in_gb * 1000 

def get_emissions_for_vm_per_month(vm_sku, disk_sku, data_out_in_gb, region='AU East'):
    return get_emissions_for_vm_sku_in_kg(vm_sku, region) + get_emissions_for_vm_disk_in_kg(disk_sku, region) + get_emissions_for_vm_network(data_out_in_gb, 'All')

# Exercise functions

d2_v2_emissions =  get_emissions_for_vm_per_month('D2 v2/DS2 v2', 'S10 Disks', 50)
print(f'VM: {d2_v2_emissions} kg in CO2Eq Emissions/Month')

d3_v2_emissions =  get_emissions_for_vm_per_month('D3 v2/DS3 v2', 'S10 Disks', 50)
print(f'VM: {d3_v2_emissions} kg in CO2Eq Emissions/Month')

f2s_v2_emissions =  get_emissions_for_vm_per_month('F2s v2', 'S10 Disks', 50)
print(f'VM: {f2s_v2_emissions} kg in CO2Eq Emissions/Month')

e2_v4_emissions =  get_emissions_for_vm_per_month('E2s v4', 'S10 Disks', 50)
print(f'VM: {e2_v4_emissions} kg in CO2Eq Emissions/Month')

# AKS - Related Emissions

def get_emissions_for_aks_in_kg():
    return az_co2_ratecard[(az_co2_ratecard.MeterCategoryName == 'Azure Kubernetes Service') & (az_co2_ratecard.MeterName == 'Standard Uptime SLA')].iloc[0].EmissionsPerUnitResourceUsage * hrs_per_month * 1000

# Generate CloudFit Emissions for test data

## Azure VMs

vm_skus = ['E2s v4', 'F2s v2', 'D3 v2/DS3 v2', 'D2 v2/DS2 v2']
#regions = ['West US', 'East US', 'North Europe', 'West Europe', 'Southeast Asia', 'East Asia', 'Australia East', 'Australia Central']
# East Asia, Southeast Asia regions are missing
regions = ['US West', 'US East', 'EU North', 'EU West','AU East', 'AU Central']

# Compute emissions data for all regions and skus defined
emissions_data = []
for region in regions:
    for sku in vm_skus:
        emissions_data.append([region, sku, get_emissions_for_vm_per_month(sku, 'S10 Disks', 50, region)])

## Workload Units

We use the Workload Units to represent a single workload/app. For VMs, 1 Workload Unit = 1 VM instance.

emissions_df = pd.DataFrame(emissions_data, columns=['Region', 'VM SKU', 'Emissions per month(kg)'])
emissions_df['Workload Units = 3'] = emissions_df['Emissions per month(kg)'] * 3
emissions_df['Workload Units = 5'] = emissions_df['Emissions per month(kg)'] * 5

emissions_df

## AKS 

With AKS - we will introduce a density factor ```density_coeff```. This is effectively a multiplier which reduces the number of nodes required due to the assumed consolidation that will occur if we migrate VM workload to containers.

With AKS 1 Workload Unit = 1 VM instance * ```density_coeff```

density_coeff = 0.5

emissions_df['AKS Workload Units = 3'] = (emissions_df['Emissions per month(kg)'] * 3 * density_coeff) + get_emissions_for_aks_in_kg()
emissions_df['AKS Workload Units = 5'] = (emissions_df['Emissions per month(kg)'] * 5 * density_coeff) + get_emissions_for_aks_in_kg()

emissions_df
