==> 06-co2signal.ipynb <==
pip install requests    

pip show requests

# How to make http request
import requests as req
apiKey = ''
r = req.get('https://api.co2signal.com/v1/latest?countryCode=US-FLA-FMPP', headers={'auth-token': apiKey})
print (r.text)

# Read CSV inputs  
import pandas as pd
df = pd.read_csv('co2signalzones\co2emissionregions.csv')
df.head(10)

# Read CSV for country code inputs  
import pandas as pd
df = pd.read_csv('co2signalzones\co2emissionregions.csv', usecols=[1])
df

#Iterating the rows of a DataFrame into http request
# https://stackoverflow.com/questions/16476924/how-to-iterate-over-rows-in-a-dataframe-in-pandas 
import time
import requests as req
import json
for index, row in df.iterrows():
    r = req.get('https://api.co2signal.com/v1/latest?countryCode='+row[1], headers={'auth-token': apiKey})
    print (r.text)
    json_data = json.loads(r.text)
    print (json_data)
    time.sleep(3) #Sleep for 3 seconds

json_data['data']


dataList =[]
for index, row in df.iterrows():
    r = req.get('https://api.co2signal.com/v1/latest?countryCode='+row[1], headers={'auth-token': apiKey})
    dataList.append(r)
    print (r.text)
    time.sleep(3) #Sleep for 3 seconds

with open('co2signalzones\co2emissionoutput.csv', 'w') as f:
    writer = csv.writer(f)
    writer.writerows(dataList)
    #df = pd.read_json(r.text)
    #df.to_csv('co2signalzones\co2emissionoutput.csv', mode='a')
